requirements.txt
Django==4.2.7
djangorestframework==3.14.0
requests==2.31.0
python-dotenv==1.0.0
pytest>=8.2,<9
pytest-django==4.5.2